+++ src/preprocessing/requirements.txt
@@ -4,3 +4,4 @@ wheel==0.34.2
 ../../../audiolib/dist/audiolib-0.1.0.tar.gz
 fsspec==0.8.0
 gcsfs==0.6.2
+-e .
